stream_uri_manager/requirements.txt
awscrt
awsiotsdk
dotenv
python-dotenv
